Add character UI

+++ RavenorAdventures/Assets/Script/Components System/Components/HealthDamageSystem/CPN_Data_HealthHandler.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public interface CPN_Data_HealthHandler
 {
     public float MaxHealth();
-    public float MaxArmor();
+    public int MaxArmor();
 
     public int Defense();
 }
